feat: add meta and define member section
