+++ app/src/components/__tests__/AddMemberCard.spec.ts
@@ -1,6 +1,6 @@
 import { it, describe, expect } from 'vitest'
 import { mount } from '@vue/test-utils'
-import AddMemberCard from '../AddMemberCard.vue'
+import AddMemberCard from '../sections/SingleTeamView/Team/AddMemberCard.vue'
 
 import { PlusCircleIcon } from '@heroicons/vue/24/outline'
 describe('AddMemberCard.vue', () => {
